remove retry logic, pull error state subscription into hook
