+++ assets/js/utils/errorHandler.tsx
@@ -9,12 +9,6 @@ export interface ErrorHandlingOptions {
   logError?: boolean;
   /** Custom error handler function */
   onError?: (error: Error | Response) => void;
-  /** Whether to retry the operation on failure */
-  retry?: boolean;
-  /** Number of retry attempts */
-  retryAttempts?: number;
-  /** Delay between retries in milliseconds */
-  retryDelay?: number;
 }
 
 export interface ErrorState {
@@ -42,11 +36,7 @@ let errorState: ErrorState | null = null;
 let errorListeners: ((error: ErrorState | null) => void)[] = [];
 let errorCount = 0;
 let lastErrorTime = 0;
-
-// Retry attempt defaults
 const ERROR_COOLDOWN_MS = 5000; // Don't show multiple errors within 5 seconds
-const RETRY_DELAY_DEFAULT_MS = 1000;
-const RETRY_ATTEMPTS_DEFAULT = 3;
 
 /**
  * Subscribes the error state by passing in a listener function to be called on state channges.
@@ -56,9 +46,6 @@ export const subscribeToErrorState = (
   listener: (error: ErrorState | null) => void,
 ) => {
   errorListeners.push(listener);
-  return () => {
-    errorListeners = errorListeners.filter((l) => l !== listener);
-  };
 };
 
 /**
@@ -71,7 +58,7 @@ export const unsubscribeFromErrorState = (
 };
 
 /**
- * Wrapper for operations to handle errors and either retry, fail silently, or display an error modal.
+ * Wrapper for operations to handle errors and either fail silently or display an error modal.
  * Defaults to displaying the error modal on first failure
  * @param asyncFunction - the function that should be tried
  * @param options - ErrorHandlingOptions
@@ -81,55 +68,29 @@ export const withErrorHandling = <T extends any[], R>(
   options: ErrorHandlingOptions = {},
 ) => {
   return async (...args: T): Promise<R | null> => {
-    const {
-      showErrorModal = true,
-      logError = true,
-      onError,
-      retry = false,
-      retryAttempts = RETRY_ATTEMPTS_DEFAULT,
-      retryDelay = RETRY_DELAY_DEFAULT_MS,
-    } = options;
+    const { showErrorModal = true, logError = true, onError } = options;
 
-    let attempts = 0;
-
-    while (attempts <= (retry ? retryAttempts : 0)) {
-      try {
-        const result = await asyncFunction(...args);
-        return result;
-      } catch (error) {
-        attempts++;
-
-        if (logError) {
-          console.error(
-            `Error in ${asyncFunction.name} (attempt ${attempts}):`,
-            error,
-          );
-        }
-
-        // If there's no handling specified and it's a session expiration error,
-        // then we want to refresh the page after displaying the error
-        if (!onError && isSessionExpirationError(error)) {
-          options.onError = () => {
-            setTimeout(() => window.location.reload(), 2000);
-          };
-        }
-
-        // If this is the last attempt or we're not retrying, handle the error
-        if (attempts > (retry ? retryAttempts : 0)) {
-          if (showErrorModal) {
-            displayErrorModal(error as Error | Response, options);
-          }
-          return null;
-        } else {
-          // Wait before retrying
-          await new Promise((resolve) =>
-            setTimeout(resolve, retryDelay * attempts),
-          );
-        }
+    try {
+      const result = await asyncFunction(...args);
+      return result;
+    } catch (error) {
+      if (logError) {
+        console.error(`Error in ${asyncFunction.name}:`, error);
       }
-    }
 
-    return null;
+      // If there's no handling specified and it's a session expiration error,
+      // then we want to refresh the page after displaying the error
+      if (!onError && isSessionExpirationError(error)) {
+        options.onError = () => {
+          setTimeout(() => window.location.reload(), 2000);
+        };
+      }
+
+      if (showErrorModal) {
+        displayErrorModal(error as Error | Response, options);
+      }
+      return null;
+    }
   };
 };
